+++ src/synack/__init__.py
@@ -3,7 +3,7 @@
 from ._handler import Handler
 from .plugins.api import Api
 from .plugins.auth import Auth
-from .plugins.db import Db
+from .db import Db
 from .plugins.debug import Debug
 from .plugins.missions import Missions
 from .plugins.notifications import Notifications
