+++ src/synack/db/models/__init__.py
@@ -1,0 +1,3 @@
+from .target import Target
+from .config import Config
+from .category import Category
