+++ src/synack/db/models/category.py
@@ -1,0 +1,18 @@
+"""db/models/category.py
+
+Database Model for the Category items
+"""
+
+import sqlalchemy as sa
+from sqlalchemy.orm import declarative_base
+
+
+Base = declarative_base()
+
+
+class Category(Base):
+    __tablename__ = 'categories'
+    id = sa.Column(sa.Integer, autoincrement=True, primary_key=True)
+    category_name = sa.Column(sa.VARCHAR(100))
+    passed_pracical = sa.Column(sa.BOOLEAN, default=False)
+    passed_written = sa.Column(sa.BOOLEAN, default=False)
